test(triggerrun): fold signal storm repro test into controller_test.go
Consolidates the standalone repro file into the package's existing test
file rather than keeping it separate.
